fix the net_detail_view tests
I forgot to use the objects.create parameters function. This way the object
never got added to the database, therefore it never showed in the
content.response. And since I use the admin interface when testing manually,
which creates the objects slightly different and correctly the error never
showed up when doing it by hand.